Add thread_local wrappers + fix unused warning
