+++ test/common/doctest.h
@@ -468,6 +468,43 @@ namespace doctest {
 
 DOCTEST_INTERFACE extern bool is_running_in_test;
 
+#if _MSC_VER
+// Change inspired by oneTBB:
+// Due to race between exiting the process and starting of a new detached thread in Windows, thread
+// local variables, which constructors or destructors have calls to runtime functions (e.g. free()
+// function) can cause access violation since TBB along with runtime library may have been unloaded
+// by the time these variables are constructed or destroyed. The workaround is do not call constructors
+// and/or destructors for such variables if they are not used.
+#include <type_traits>
+template <typename T>
+struct doctest_thread_local_wrapper {
+    doctest_thread_local_wrapper() {
+        // Default definition is ill-formed in case of non-trivial type T.
+    }
+    T& get() {
+        if( !initialized ) {
+            new(&value) T;
+            initialized = true;
+        }
+        return value;
+    }
+    ~doctest_thread_local_wrapper() {
+        if( initialized )
+            value.~T();
+    }
+private:
+    union { T value; };
+    bool initialized = false;
+};
+#else  // _MSC_VER
+template <typename T>
+struct doctest_thread_local_wrapper {
+    T& get() { return value; }
+private:
+    T value{};
+};
+#endif // _MSC_VER
+
 // A 24 byte string class (can be as small as 17 for x64 and 13 for x86) that can hold strings with length
 // of up to 23 chars on the stack before going on the heap - the last byte of the buffer is used for:
 // - "is small" bit - the highest bit - if "0" then it is small - otherwise its "1" (128)
@@ -3018,9 +3055,10 @@ namespace detail {
         return oss.str().c_str();
     }
 
-    DOCTEST_THREAD_LOCAL std::ostringstream g_oss; // NOLINT(cert-err58-cpp)
+    DOCTEST_THREAD_LOCAL doctest_thread_local_wrapper<std::ostringstream> wrapped_g_oss; // NOLINT(cert-err58-cpp)
 
     std::ostream* getTlsOss() {
+        auto& g_oss = wrapped_g_oss.get();
         g_oss.clear(); // there shouldn't be anything worth clearing in the flags
         g_oss.str(""); // the slow way of resetting a string stream
         //g_oss.seekp(0); // optimal reset - as seen here: https://stackoverflow.com/a/624291/3162383
@@ -3029,7 +3067,7 @@ namespace detail {
 
     String getTlsOssResult() {
         //g_oss << std::ends; // needed - as shown here: https://stackoverflow.com/a/624291/3162383
-        return g_oss.str().c_str();
+        return wrapped_g_oss.get().str().c_str();
     }
 
 #ifndef DOCTEST_CONFIG_DISABLE
@@ -3991,7 +4029,8 @@ namespace {
         return 0;
     }
 
-    int dumy_init_console_colors = colors_init();
+    // Change inspired by oneTBB : add volatile to get rid of "unused variable warning" + fix typo
+    volatile int dummy_init_console_colors = colors_init();
 #endif // DOCTEST_CONFIG_COLORS_WINDOWS
 
     DOCTEST_CLANG_SUPPRESS_WARNING_WITH_PUSH("-Wdeprecated-declarations")
@@ -4185,10 +4224,10 @@ namespace detail {
     void toStream(std::ostream* s, int long long in) { *s << in; }
     void toStream(std::ostream* s, int long long unsigned in) { *s << in; }
 
-    DOCTEST_THREAD_LOCAL std::vector<IContextScope*> g_infoContexts; // for logging with INFO()
+    DOCTEST_THREAD_LOCAL doctest_thread_local_wrapper<std::vector<IContextScope*>> wrapped_g_infoContexts; // for logging with INFO()
 
     ContextScopeBase::ContextScopeBase() {
-        g_infoContexts.push_back(this);
+        wrapped_g_infoContexts.get().push_back(this);
     }
 
     // Change inspired by oneTBB : fix copy elision absence case
@@ -4197,7 +4236,7 @@ namespace detail {
             other.destroy();
         }
         other.need_to_destroy = false;
-        g_infoContexts.push_back(this);
+        wrapped_g_infoContexts.get().push_back(this);
     }
 
     DOCTEST_MSVC_SUPPRESS_WARNING_WITH_PUSH(4996) // std::uncaught_exception is deprecated in C++17	
@@ -4217,7 +4256,7 @@ namespace detail {
             this->stringify(&s);
             g_cs->stringifiedContexts.push_back(s.str().c_str());
         }
-        g_infoContexts.pop_back();
+        wrapped_g_infoContexts.get().pop_back();
     }
 
     DOCTEST_CLANG_SUPPRESS_WARNING_POP	
@@ -6013,19 +6052,19 @@ namespace {
 #ifdef DOCTEST_PLATFORM_WINDOWS
     struct DebugOutputWindowReporter : public ConsoleReporter
     {
-        DOCTEST_THREAD_LOCAL static std::ostringstream oss;
+        DOCTEST_THREAD_LOCAL static doctest_thread_local_wrapper<std::ostringstream> wrapped_oss;
 
         DebugOutputWindowReporter(const ContextOptions& co)
-                : ConsoleReporter(co, oss) {}
+                : ConsoleReporter(co, wrapped_oss.get()) {}
 
 #define DOCTEST_DEBUG_OUTPUT_REPORTER_OVERRIDE(func, type, arg)                                    \
     void func(type arg) override {                                                                 \
         bool with_col = g_no_colors;                                                               \
         g_no_colors   = false;                                                                     \
         ConsoleReporter::func(arg);                                                                \
-        if(oss.tellp() != std::streampos{}) {                                                      \
-            DOCTEST_OUTPUT_DEBUG_STRING(oss.str().c_str());                                        \
-            oss.str("");                                                                           \
+        if(wrapped_oss.get().tellp() != std::streampos{}) {                                                      \
+            DOCTEST_OUTPUT_DEBUG_STRING(wrapped_oss.get().str().c_str());                                        \
+            wrapped_oss.get().str("");                                                                           \
         }                                                                                          \
         g_no_colors = with_col;                                                                    \
     }
@@ -6043,7 +6082,7 @@ namespace {
         DOCTEST_DEBUG_OUTPUT_REPORTER_OVERRIDE(test_case_skipped, const TestCaseData&, in)
     };
 
-    DOCTEST_THREAD_LOCAL std::ostringstream DebugOutputWindowReporter::oss;
+    DOCTEST_THREAD_LOCAL doctest_thread_local_wrapper<std::ostringstream> DebugOutputWindowReporter::wrapped_oss;
 #endif // DOCTEST_PLATFORM_WINDOWS
 
     // the implementation of parseOption()
@@ -6600,9 +6639,9 @@ DOCTEST_MSVC_SUPPRESS_WARNING_POP
 
 IReporter::~IReporter() = default;
 
-int IReporter::get_num_active_contexts() { return detail::g_infoContexts.size(); }
+int IReporter::get_num_active_contexts() { return detail::wrapped_g_infoContexts.get().size(); }
 const IContextScope* const* IReporter::get_active_contexts() {
-    return get_num_active_contexts() ? &detail::g_infoContexts[0] : nullptr;
+    return get_num_active_contexts() ? &detail::wrapped_g_infoContexts.get()[0] : nullptr;
 }
 
 int IReporter::get_num_stringified_contexts() { return detail::g_cs->stringifiedContexts.size(); }
