+++ test/common/test.h
@@ -29,42 +29,6 @@
 #define TBB_TEST_THROW(x) FAIL("Exceptions are disabled")
 #endif
 
-#if _MSC_VER
-// Due to race between exiting the process and starting of a new detached thread in Windows, thread
-// local variables, which constructors or destructors have calls to runtime functions (e.g. free()
-// function) can cause access violation since TBB along with runtime library may have been unloaded
-// by the time these variables are constructed or destroyed. The workaround is do not call constructors
-// and/or destructors for such variables if they are not used.
-#include <type_traits>
-template <typename T>
-struct doctest_thread_local_wrapper {
-    doctest_thread_local_wrapper() {
-        // Default definition is ill-formed in case of non-trivial type T.
-    }
-    T& get() {
-        if( !initialized ) {
-            new(&value) T;
-            initialized = true;
-        }
-        return value;
-    }
-    ~doctest_thread_local_wrapper() {
-        if( initialized )
-            value.~T();
-    }
-private:
-    union { T value; };
-    bool initialized = false;
-};
-#else  // _MSC_VER
-template <typename T>
-struct doctest_thread_local_wrapper {
-    T& get() { return value; }
-private:
-    T value{};
-};
-#endif // _MSC_VER
-
 #include "doctest.h"
 
 #define CHECK_FAST(x) do { if (!(x)) { CHECK(false); } } while((void)0, 0)
